+++ esp-idf-svc/CHANGELOG.md
@@ -39,6 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `use_serde` feature, which enables the `use_serde` feature of `embedded-svc` crate, allowing to deserialize configuration structs.
 - OTA: Allow specifying image size to speed up erase
 - Bluetooth: New methods `EspBleGap::start_scanning` and `EspBleGap::stop_scanning`
+- Bluetooth Classic: Added Serial Port Profile, `spp`
+- New example, `bt_spp_acceptor` to demonstrate usage of bt classic spp profile
 - New example, `bt_ble_gap_scanner` to demonstrate usage of added ble scanning methods
 - New example, `mdns_advertise` to demonstrate mDNS service advertisement
 - NVS: Implemented `RawHandle` for `EspNvs<NvsDefault>`